requirements.txt
flask>=2.0.0
requests>=2.28.0
pysocks>=1.7.1
stem>=1.8.0
click>=8.0.0
pyyaml>=6.0
